+++ collie/cross_validation.py
@@ -141,7 +141,7 @@ def stratified_split(interactions: BaseInteractions,
                      test_p: float = 0.2,
                      processes: int = -1,
                      seed: Optional[int] = None,
-                     ignore_warning: bool = False) -> Tuple[BaseInteractions, ...]:
+                     force_split: bool = False) -> Tuple[BaseInteractions, ...]:
     """
     Split an ``Interactions`` instance into train, validate, and test datasets in a stratified
     manner such that each user appears at least once in each of the datasets.
@@ -177,8 +177,11 @@ def stratified_split(interactions: BaseInteractions,
         will be used
     seed: int
         Random seed for splits
-    ignore_warning: bool
-        Ignore warning when users have only one interaction, force those users into the training set
+    force_split: bool
+        Ignore error raised when a user in the dataset has only a single interaction. Normally,
+        a ``ValueError`` is raised when this occurs. When ``force_split=True``, however,
+        users with a single interaction will be placed in the training set and an error will NOT be
+        raised
 
     Returns
     -------
@@ -214,14 +217,14 @@ def stratified_split(interactions: BaseInteractions,
                                     test_p=test_p,
                                     processes=processes,
                                     seed=seed,
-                                    ignore_warning=ignore_warning)
+                                    force_split=force_split)
 
     if val_p > 0:
         train, validate = _stratified_split(interactions=train,
                                             test_p=val_p / (1 - test_p),
                                             processes=processes,
                                             seed=seed,
-                                            ignore_warning=ignore_warning)
+                                            force_split=force_split)
 
         return train, validate, test
     else:
@@ -232,7 +235,7 @@ def _stratified_split(interactions: BaseInteractions,
                       test_p: float,
                       processes: int,
                       seed: int,
-                      ignore_warning: bool) -> Tuple[Interactions, Interactions]:
+                      force_split: bool) -> Tuple[Interactions, Interactions]:
     users = interactions.mat.row
     unique_users = set(users)
 
@@ -249,7 +252,7 @@ def _stratified_split(interactions: BaseInteractions,
             _stratified_split_parallel_worker(idxs_to_split=all_idxs_for_users_dict[user],
                                               test_p=test_p,
                                               seed=(seed + user),
-                                              ignore_warning=ignore_warning)
+                                              force_split=force_split)
             for user in unique_users
         ]
     else:
@@ -261,7 +264,7 @@ def _stratified_split(interactions: BaseInteractions,
             delayed(_stratified_split_parallel_worker)(all_idxs_for_users_dict[user],
                                                        test_p,
                                                        seed + user,
-                                                       ignore_warning)
+                                                       force_split)
             for user in unique_users
         )
 
@@ -281,7 +284,7 @@ def _stratified_split(interactions: BaseInteractions,
 def _stratified_split_parallel_worker(idxs_to_split: Iterable[Any],
                                       test_p: float,
                                       seed: int,
-                                      ignore_warning: bool) -> np.array:
+                                      force_split: bool) -> np.array:
     try:
         _, test_idxs = train_test_split(idxs_to_split,
                                         test_size=test_p,
@@ -290,10 +293,11 @@ def _stratified_split_parallel_worker(idxs_to_split: Iterable[Any],
                                         stratify=np.ones_like(idxs_to_split))
     except ValueError as ve:
         if 'the resulting train set will be empty' in str(ve):
-            if ignore_warning is False:
+            if force_split is False:
                 raise ValueError(
-                    'Unable to straify split on users. The ``interactions`` object contains users '
-                    'with only one interaction, either set ``ignore_warning = True` or run '
+                    'Unable to stratify split on users. The ``interactions`` object contains users '
+                    'with a single interaction, either set ``force_split = True`` to put all users '
+                    'with a single interaction in the training set or run '
                     '``collie.utils.remove_users_with_fewer_than_n_interactions`` first.'
                 )
             else: