+++ collie/model/hybrid_pretrained_matrix_factorization.py
@@ -279,6 +279,7 @@ class HybridPretrainedModel(BasePipeline):
 
     def _get_user_embeddings(self) -> torch.tensor:
         """Get user embeddings on device."""
+        # TODO: update this to get the embeddings post-MLP
         return self.embeddings[0].weight.data
 
     def freeze_embeddings(self) -> None: